Enhance contract deployment in integration tests (#9318)
This pr improves contract deployment during tests.

Previously on deploy we `sendTopLevelCreate`
and used the mirrorResult from 
`contractExecutionService.processCall(serviceParameters);` which returns the contract runtime code.

With this new change: instead of calling `contractExecutionService.processCall` we are using a new utility method which quickly extracts 
the runtime bytecode from the `rawTransaction.getData() `which is the init bytecode.

Using this approach we skip the need for calling `contractExecutionService.processCall` in 99% of the test cases. For some corner cases we still use the old approach with the `contractExecutionService.processCall` .

Couple of possible scenarios explored but **not applied in this PR**.

 - Find a way to even skip persisting the contract to db on every call. Did not find an straightorward way to do that.

 - Add a before each method in every test suite which with the help of a flag will deploy the contract only in the first before 
each.
and the rest of the before each executions will do nothing. This will mean we only deploy the contract only once during a full test class execution. Unfortunately this approach led to 0 peformance gain.


This PR modifies:
Add a utility class RuntimeBytecodeExtractor - which has a method that returns the runtime bytecode and takes init bytecode as a parameter.

Adds a test class RuntimeBytecodeExtractorTest - which tests the utility class method with different scenarios.

Modified TestWeb3jService - adds a new flag - which when enabled uses `contractExecutionService.processCall` for contract deployment. By default it will use the approach with using the new utility class.

Modified couple of corner test cases to preserve contract deployment with `contractExecutionService.processCall`

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>