Add negative scenarios for historical tests (#9345)
This PR adds some new test related to the historical scenarios. They verify that the correct data is fetched having a particular historical timestamp. The tests are located in `ContractCallServiceHistoricalNegativeTest`.

A new abstract class is added - `AbstractContractCallServiceHistoricalTest`. It contains some common methods related to the historical data persist.

The methods in `AbstractContractCallServiceHistoricalTest` and `AbstractContractCallServiceTest` are named with <entity_type>persist instead of persist<entity_type> in order to use the IDE's auto suggest easily. The second approach would list too much methods as they all start with "p" and this makes the search harder, so I renamed some of the old methods.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>