Add `CustomCallOperation` (#9538)
In services there is a `CustomCallOperation` class. In there the cost of the operation is calculated as if the address is always `warm == false`. In mirror node we use the Besu's `CallOperation` which determines the cost based on whether the address is warm or not. This results in a lower cost for warm addresses and hence the expected gas estimation is lower than the correct value. The added `CustomCallOperation` in this PR is really simplified and ignores some checks from the original class in services. The ignored checks are related to missing addresses. The reasons for that are the following:
1. We have similar checks elsewhere.
2. The lazy creation is enabled in mirror node.
3. The effort to copy the exact implementation in services would be a lot and when we complete the reusable services epic soon the code in this PR will be removed anyway.

After the changes in this PR the returned gas estimation is really close to the padded value in the issue below.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>