Implement `TokenRelationshipReadableKVState` (#9582)
As part of the implementation for the reusable services we need to implement TokenRelationshipReadableKVState. It reads a token account association and converts it to TokenRelation PBJ type.

The TokenRelation class is a copy of the same class from PBJ. The difference is in the balance field - in this custom implementation it is replaced with a Supplier object in order to use memoization. The balance needs to be calculated with a DB query for the historical cases as we don't have this value directly.

We need to read the token type, the account entity and the token balances in order to convert to a TokenRelation object. The token type is read by introducing a new DB query that returns only the required column instead of using the existing queries that return the full Token object.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>