fix: Fix contract deployment with value (#9668)
This PR fixes contract deployment when the contract deployment is made with value.

This PR modifies :
Modified ContractCallRequest - change the 'to' validation to check if the call is for contract deployment

BytecodeUtils - Added a new method isInitBytecode which checks if a given data is an init bytecode - needed in ContractCallRequest validation

ContractControllerTest - Added new test and modified a few existing ones based on the new changes to the validation of the to field.

Modified ContractCallAddressThisTest - added tests for deployment of payable contract with and without value
Modified ContractCallServiceERCTokenModificationFunctionsTest - added tests for deployment of non payable contract with and without value.

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>