Create `State` implementation with registering all of the specific `ReadableKVStates` (#9700)
This PR adds a MirrorNodeState class that implements the State interface. The class registers all of needed readable and writable states. Each service (TokenService, ContractService or FileService) needs a different set of states in order to function properly. The readable states are initialised only once and are singleton instances and will be used by all of the threads. They are the implementations of all of the ReadableKVStates. For the writable states we need only 1 facade implementation since no actual changes will be persisted from the memory. The writable state uses a readable state for the read operations and the modifications are kept in memory and discarded at the end of the transaction execution.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>