Cherry pick 9868 add additional coverage for token airdrop acceptance tests (0.119) (#9885)
Add additional coverage for token airdrops tests (#9868)

This PR adds additional coverage for fungible and non-fungible token airdrops in cases when the account is not associated to the token and claims the airdrop. The association should happen upon claim. Also token airdrops tests are extracted in separate test scenarios for cleaner overview

---------

Signed-off-by: Nikolay Nikolov <nikolay.nikolov@limechain.tech>