Add file id and init bytecode in `ContractCallContext` (#9799)
During the implementation of the reusable services integration it turns out that now the contracts are deployed in 2 steps:

    The initcode is uploaded and saved as a file using a FileCreateTransactionBody.
    The returned file id from step 1 is then passed to a ContractCreateTransactionBody.
    Each step performs a separate transaction.
    For step 2 even if we pass the correct file id, since the mirror node data is readonly, the FileReadableKVState is not able to populate the contract's bytecode from the DB since it was never explicitly persisted in the DB.

This PR add a new field in the ContractCallContext that preserve the file id and the contract's bytecode. The new field is called "file". If it is available when we call FileReadableKVState#readFromDataSource then we return the file from the context with priority.

This field might need to be cleared up at some point but this will probably added later when the correct place is found during testing.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>