Introduce `gasUnit` property to allow greater gas per second throttle (#9994)
The current bucket4j implementation does not allow a greater consumption than 1 token per nanosecond. Because of this the current max gas per second is capped at 1_000_000_000 gas.

This PR adds a new property called hedera.mirror.web3.throttle.gasUnit. This is a multiplier that allows a mapping between the token and the gas that will be consumed in the bucket. For example, if the value of the property is set to 10, this means that 1 token in the bucket will correspond to 10 gas. This will make the max supported gas limit per second to be equal to 10_000_000_000. Please note that there might be some small precision loss due to number rounding.

The implementation change is that when a request is made, the passed gas limit is divided by the configured gasUnit property value (by default it is set to 1). After the transaction processing is finished, the remaining (unused) gas is then returned back to the bucket and is available to be used by another transaction.
Example: a request is made with gasLimit set to 1_000_000. The gasUnit is set to 10. This means that 100_000 tokens will be reserved in the gas throttle bucket. The transaction finishes and the actual gas used is 800_000. This means that there are 200_000 gas remaining. Using the gasUnit multiplier, 200_000 / gasUnit = 20_000 gas will be returned back to the bucket and can be used by another transaction.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>