Fix some `INSUFFICIENT_PAYER_BALANCE` errors (#10124)
After PR #10063 was merged, some of the test that were previously successful, now are failing with INSUFFICIENT_PAYER_BALANCE error. The reason for that was that before the PR, the tests were false positive. After the PR, the hidden errors are now uncovered. This PR moves the treasury account and balance creation in AbstractContractCallServiceTest#setup only. This PR is focused on the errors in ContractCallServicePrecompileModificationTest.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>