Remove file upload on contract create and use `maxSignedTxnSize` property (#10218)
After the hedera app dependency bump to 0.58.x we now have available the `maxSignedTxnSize` property. This allows us to set it to a value higher than 6KB that was the default one before. The configured callData size in the requests is currently 128 KB. The size of the call data + the additional transaction fields and signing increase the overall transaction size a little (about 50-200 bytes depending on the request), so value of `maxSignedTxnSize` is now set to 129KB to ensure that there is enough buffer and that the transaction will not be rejected in hedera app.

Setting this property allows us to:
1. Remove the checks in `TransactionExecutionService` for transaction size.
2. Remove the file upload logic for contract create when the transaction is bigger that 6KB.
3. Remove the `file` field in the `ContractCallContext` that saved the content of the uploaded file between the separate transaction executions (file uplaod & contract create).
4. Clean the `FileReadableKVState` from using the `file` field from the `ContractCallContext`.

This makes the overall code a lot cleaner and easier to understand.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>