Fix the ERC historical tests (#10251)
This PR fixes the setup for the ERC historical tests. Also, the timestamp from the context is now fetched differently:

    If there is timestamp set directly -> return this timestamp.
    If there isn't -> return the timestamp from the consensus end timestamp of the set record file in the context.
    The above points are valid only if we are in a historical context. Otherwise -> return empty optional.

Note: Changing the timestamp fetch logic now enables us to read historical records correctly. However, there is a cache mechanism in ReadableKVStateBase and when multiple test suites are run it gets polluted - some tests fail when all tests are run and pass when run separately. This is a known issue but the change from this PR seems to make it appear more frequently. This is not a regression.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>
Signed-off-by: Harsh Sawarkar <harshsawarkar111@gmail.com>