Fix OpcodeServiceTest for modularized (#10247)
This PR fixes 80+ tests from OpcodeServiceTest
Fixed ContractCallDynamicCallsTest- mintTokenGetTotalSupplyAndBalanceOfTreasury which is also used in OpcodeServiceTest.
PR is checkout from #10217 since it contains same setup in as in NestedCalls tests.
Must wait for #10217 to be merged first.
The only test that is not passing is callWithDifferentCombinationsOfTracerOptions with the exception of the test that set storage to true in OpcodeTracer().

   public Map<UInt256, UInt256> getUpdatedStorage() {
        throw new UnsupportedOperationException("Storage changes are managed by the EvmFrameState");
    }

fails in upstream code as not supported.

A few tests in updateTokenKeysAndGetUpdatedTokenKeyForNFT with CONSENSUS_GAS_EXHAUSTED but this will be fixed in another PR.

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>
Signed-off-by: Harsh Sawarkar <harshsawarkar111@gmail.com>