Handle exception throwing in binary gas estimator when performing estimate gas (#10258)
This PR adds an exception handling during estimate gas using transaction executor. Using the modularized services when performing a contract call in binary gas estimator with the mid value and if the call fails it throws an exception instead of result with status failed. That results in a failing in the test. 

- Added a method that performs a safe contract call with try-catch mechanism
- Added additional check for 'err' if the contract result is null
- added logic to differentiate between the two flows for modularizedServices

This PR also fixes 1 test in contractCallSErviceTest

---------

Signed-off-by: Nikolay Nikolov <nikolay.nikolov@limechain.tech>