Fix last failing test in precompile modification tests (#10313)
This PR fixes the last failing test against modularized code -> notExistingPrecompileCall

According to docs - > When you attempt to make a call to a non-existing precompile addres the behavior is designed to be equivalent to Ethereum for compatibility reasons. The call will execute successfully.

Unfortunately mono does not cover it so it still reverts. The test will assert based on the modularized flag until full transition to modularized execution

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>