Mirror node state scoped cache implementation (#10306)
With this PR the `ReadableKVStateBase` and the `WritableKVStateBase` classes from the hedera app dependency have been replaced with custom implementations. The main difference is that the base classes have private maps that cache the read/modified entries and these classes are singletons that are shared between all transactions. This lead to inconsistencies such as:
1. Transaction T1 reads account A against the latest block and puts it in the cache.
2. Transaction T2 reads account A against a historical block where account A does not exist yet and since the cache is shared, A is returned instead of null.
3. Other similar cases where and entry needs to be returned, but the cache keeps its value as null.
The case for the writable cache is similar to the readable one.

This PR replaces the private cache maps with scoped value maps. They are now located in the `ContractCallContext` to ensure that each transaction has its own small cache that would not be affected by other transactions. Also, this cache will be valid only in the current context, so at the end of the scope the cache will be cleared automatically and no additional action is needed. A special case is the gas estimation - in there we reset the modifications cache to ensure a fresh start. The read cache is not currently cleared as it would contain only the test setup and this would not change between the different iterations. This way we skip some repository calls as well for some optimisation.

The read and the write cache in the `ContractCallContext` are kept as map of maps - the key for the parent map is the state key, e.g. "ACCOUNTS", "TOKENS", etc. The value of the parent map is a map itself that keeps the actual cache as key-value pairs the same way as it was before in the base classes.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>