Fix the rest `ContractCallDynamicCallsTest` tests (#10381)
This PR fixes the setup for the ContractCallDynamicCallsTest tests when the modularizedServices flag is set to true. Also, the TransactionExecutionService is modified in case of a failure in an opcode trace scenario to return a failed result instead of throwing an exception directly. The reason for this is that we need to be able to return the opcode list from the context before throwing an exception.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>